pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redteam"
version = "1.0.0"
description = "AI-Driven Red Team Simulation Framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "RedTeam Framework", email = "redteam@security.local"}
]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "python-nmap>=0.7.1",
    "pyyaml>=6.0",
    "requests>=2.31",
    "jinja2>=3.1",
    "networkx>=3.1",
    "flask>=3.0",
    "flask-socketio>=5.3",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
]

[project.scripts]
redteam = "redteam.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["redteam*"]
